extract sources to separate files
