+++ packages/ragbits-document-search/src/ragbits/document_search/documents/document.py
@@ -5,7 +5,8 @@ from typing import Annotated
 
 from pydantic import BaseModel
 
-from ragbits.document_search.documents.sources import LocalFileSource, Source, SourceDiscriminator
+from ragbits.document_search.documents.local_file_source import LocalFileSource
+from ragbits.document_search.documents.sources import Source, SourceDiscriminator
 
 
 class DocumentType(str, Enum):
