+++ packages/ragbits-evaluate/src/ragbits/evaluate/pipelines/document_search.py
@@ -4,7 +4,7 @@ from uuid import uuid4
 from typing_extensions import Self
 
 from ragbits.document_search import DocumentSearch
-from ragbits.document_search.documents.sources import HuggingFaceSource
+from ragbits.document_search.documents.hugging_face_source import HuggingFaceSource
 from ragbits.evaluate.pipelines.base import EvaluationPipeline, EvaluationResult
